[Hitesh] added remainder line

+++ Calculator/Program.cs
@@ -31,7 +31,7 @@ namespace Calculator
             Console.WriteLine("Division of {0} & {1} is: {2} ", num1, num2,div);
 
             //remainder
-            float rem = num1 % num2;
+            int rem = num1 % num2;
             Console.WriteLine("Remainder of {0} & {1} is: {2} ", num1, num2, rem);      }
 
     }
